% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h6-package.R
\docType{package}
\name{h6-package}
\alias{h6-package}
\title{Advanced R programming, homework 6}
\description{
This package is a solution of homework 6 of R Advanced Programming course.
It consists of three functions: mode, simplify2array2 and perms.
More about the course: https://courses.ipipan.edu.pl/simple.php?page=syllabus018.html.
}

