% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{- an integer vector}
}
\value{
- the most frequently occuring value
}
\description{
This function returns the most frequently occuring value in an integer vector (mode).
  If there are more than one values satusfying this definition, it returns one of them
  (but there is no rule which one).
}

