% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{perms}
\usage{
perms(n)
}
\arguments{
\item{n}{- number of elements to permute}
}
\value{
- matrix with all permutations of 1,2,...,n sequence
}
\description{
This function returns matrix with all permutations of sequence 1,2,...,n
  where n is given as input.
}

