% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array2}
\alias{simplify2array2}
\title{simplify2array2}
\usage{
simplify2array2(x)
}
\arguments{
\item{x}{- list}
}
\value{
If input list consists of only numeric vectors of equal length then
  the return value is its array version. Otherwise input list is returned.
}
\description{
This function transforms input list into an array if it is possible.
}

